@@ -6,13 +6,13 @@ import org.springframework.context.annotation.Configuration;
 
 import java.time.LocalDate;
 import java.time.Month;
-import java.util.Arrays;
+import java.util.List;
 
 @Configuration
 public class ItemConfig {
     @Bean
     CommandLineRunner commandLineRunner(ItemRepository itemRepository) {
-        return args -> itemRepository.saveAll(Arrays.asList(
+        return args -> itemRepository.saveAll(List.of(
                 new Item(1L, "Laptop", "Lenovo", "A1", LocalDate.of(2022, Month.JANUARY, 1)),
                 new Item(2L,"Laptop", "Lenovo", "A1", LocalDate.of(2022, Month.JANUARY, 1)),
                 new Item(3L,"Laptop", "Lenovo", "A1", LocalDate.of(2022, Month.JANUARY, 1)),
